Fix a bug in identifying pops in shadow stack pass.
Fix a bug in identifying pops to instrument in the shadow stack pass.
For pops that write to LR, we instrumented them if they are followed by
a tail call in the same basic block; this is because there was no way to
identify if a pop is part of function epilogue.  However, a tail call
might not reside in the same basic block in some cases, which leaves the
frame-destroying pop uninstrumented.  Now with 24065c1, we can use the
MachineInstr::FrameDestroy flag to distinguish them.