Fix a bug in shadow stack pass.
Fix a bug in the shadow stack pass, where we took it for granted that
R12 (IP) would be free to use at every push in function prologue and
every pop in function epilogue and hence used R12 unconditionally for
our shadow stack instrumentation.  In fact, R12 might not be free in
some cases.  This commit changes the shadow stack pass to use
findFreeRegistersBefore() and findFreeRegistersAfter() utility
functions to find a free register and, if no free register is found,
resort to register spilling.